2022/finalMovieProject/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
six==1.16.0
Werkzeug==2.0.1
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.1
Flask-SocketIO==5.1.1
Flask-Session==0.4.0
Flask-Login==0.5.0
Flask-Cors==3.0.10
requests==2.22.0
psycopg2